requirements.txt
streamlit
torch
torchvision
fastai
pillow